requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
duckdb==1.1.3
effdet==0.4.1
emoji==2.14.0
etelemetry==0.3.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-vision==3.9.0
google-generativeai==0.8.3
google_search_results==2.4.2
googleapis-common-protos==1.66.0
groq==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
gspread-pandas==3.3.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
icecream==2.1.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.28
langchain-google-genai==2.0.7
langchain-groq==0.2.2
langchain-text-splitters==0.3.3
langdetect==1.0.9
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.3
layoutparser==0.3.4
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.18.4
networkx==3.2.1
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.21.0
pikepdf==9.4.2
pillow==11.0.0
portalocker==3.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
prov==2.0.1
psutil==6.1.0
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.3
Pygments==2.18.0
PyMuPDF==1.25.1
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.19
python-oxmsg==0.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
RapidFuzz==3.10.1
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scipy==1.13.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sql_metadata==2.15.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
st-gsheets-connection==0.1.0
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
timm==1.0.12
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traits==6.4.3
transformers==4.47.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.11
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==1.26.20
validators==0.34.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
